# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Clinic Digital Growth SaaS Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tokens/queue:
    get:
      operationId: getTokenQueue
      tags: [tokens]
      summary: Get today's token queue for the authenticated clinic
      security:
        - bearerAuth: []
      parameters:
        - name: date
          in: query
          schema:
            type: string
          description: Date (YYYY-MM-DD), defaults to today
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TokenWithAppointment"
  /tokens/{tokenId}/status:
    patch:
      operationId: updateTokenStatus
      tags: [tokens]
      summary: Update token status (call next, complete, skip)
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/tokenId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTokenStatusBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Token"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # Subscriptions
  /subscriptions/current:
    get:
      operationId: getCurrentSubscription
      tags: [subscriptions]
      summary: Get current subscription for the authenticated clinic
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubscriptionStatus"
        "404":
          description: No subscription
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /subscriptions/history:
    get:
      operationId: getSubscriptionHistory
      tags: [subscriptions]
      summary: Get subscription history for the authenticated clinic
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Subscription"
  /subscriptions/{clinicId}/upgrade:
    post:
      operationId: upgradeSubscription
      tags: [subscriptions]
      summary: Upgrade a clinic subscription (admin only)
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/clinicId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpgradeSubscriptionBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"

  # Payments
  /payments/create-order:
    post:
      operationId: createPaymentOrder
      tags: [payments]
      summary: Create a Razorpay payment order
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RazorpayOrder"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /payments/verify:
    post:
      operationId: verifyPayment
      tags: [payments]
      summary: Verify Razorpay payment signature
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyPaymentBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentVerificationResponse"
        "400":
          description: Invalid signature
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /payments/webhook:
    post:
      operationId: paymentWebhook
      tags: [payments]
      summary: Razorpay webhook endpoint
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /payments/history:
    get:
      operationId: listPayments
      tags: [payments]
      summary: List payments for the authenticated clinic
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Payment"

  # Invoices
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    clinicId:
      name: clinicId
      in: path
      required: true
      schema:
        type: integer
...
    tokenId:
      name: tokenId
      in: path
      required: true
      schema:
        type: integer
...
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    # Auth schemas
...
    Appointment:
      type: object
      properties:
        id:
          type: integer
        clinicId:
          type: integer
        doctorId:
          type: ["integer", "null"]
        patientName:
          type: string
        patientPhone:
          type: string
        patientEmail:
          type: ["string", "null"]
        appointmentDate:
          type: string
        timeSlot:
          type: string
        reason:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        status:
          type: string
          enum: [waiting, in_progress, completed, cancelled, rescheduled, no_show]
        rescheduledDate:
          type: ["string", "null"]
        rescheduledTimeSlot:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - clinicId
        - patientName
        - patientPhone
        - appointmentDate
        - timeSlot
        - status
        - createdAt
        - updatedAt
...
    Token:
      type: object
      properties:
        id:
          type: integer
        clinicId:
          type: integer
        appointmentId:
          type: integer
        tokenNumber:
          type: integer
        tokenDate:
          type: string
        status:
          type: string
          enum: [waiting, called, completed, skipped]
        calledAt:
          type: ["string", "null"]
          format: date-time
        completedAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - clinicId
        - appointmentId
        - tokenNumber
        - tokenDate
        - status
        - createdAt
    TokenWithAppointment:
      allOf:
        - $ref: "#/components/schemas/Token"
        - type: object
          properties:
            appointment:
              $ref: "#/components/schemas/Appointment"
    UpdateTokenStatusBody:
      type: object
      properties:
        status:
          type: string
          enum: [waiting, called, completed, skipped]
      required:
        - status

    # Subscription schemas
    Subscription:
      type: object
      properties:
        id:
          type: integer
        clinicId:
          type: integer
        plan:
          type: string
          enum: [monthly, quarterly, yearly]
        status:
          type: string
          enum: [active, expired, cancelled, pending]
        startDate:
          type: ["string", "null"]
          format: date-time
        endDate:
          type: ["string", "null"]
          format: date-time
        amount:
          type: integer
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - clinicId
        - plan
        - status
        - amount
        - createdAt
    SubscriptionStatus:
      type: object
      properties:
        isActive:
          type: boolean
        subscription:
          $ref: "#/components/schemas/Subscription"
        daysRemaining:
          type: ["integer", "null"]
        expiresAt:
          type: ["string", "null"]
          format: date-time
      required:
        - isActive
    UpgradeSubscriptionBody:
      type: object
      properties:
        plan:
          type: string
          enum: [monthly, quarterly, yearly]
        durationMonths:
          type: integer
      required:
        - plan

    # Payment schemas
    CreateOrderBody:
      type: object
      properties:
        plan:
          type: string
          enum: [monthly, quarterly, yearly]
      required:
        - plan
    RazorpayOrder:
      type: object
      properties:
        orderId:
          type: string
        amount:
          type: integer
        currency:
          type: string
        plan:
          type: string
        keyId:
          type: string
      required:
        - orderId
        - amount
        - currency
        - plan
        - keyId
    VerifyPaymentBody:
      type: object
      properties:
        razorpayOrderId:
          type: string
        razorpayPaymentId:
          type: string
        razorpaySignature:
          type: string
        plan:
          type: string
          enum: [monthly, quarterly, yearly]
      required:
        - razorpayOrderId
        - razorpayPaymentId
        - razorpaySignature
        - plan
    PaymentVerificationResponse:
      type: object
      properties:
        success:
          type: boolean
        payment:
          $ref: "#/components/schemas/Payment"
        subscription:
          $ref: "#/components/schemas/Subscription"
        invoice:
          $ref: "#/components/schemas/Invoice"
      required:
        - success
    Payment:
      type: object
      properties:
        id:
          type: integer
        clinicId:
          type: integer
        subscriptionId:
          type: ["integer", "null"]
        razorpayOrderId:
          type: ["string", "null"]
        razorpayPaymentId:
          type: ["string", "null"]
        amount:
          type: integer
        currency:
          type: string
        status:
          type: string
          enum: [pending, success, failed, refunded]
        method:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - clinicId
        - amount
        - currency
        - status
        - createdAt

    # Invoice schemas
    Invoice:
      type: object
      properties:
        id:
          type: integer
        invoiceNumber:
          type: string
        clinicId:
          type: integer
        paymentId:
          type: ["integer", "null"]
        subscriptionId:
          type: ["integer", "null"]
        amount:
          type: integer
        currency:
          type: string
        description:
          type: ["string", "null"]
        issuedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - invoiceNumber
        - clinicId
        - amount
        - currency
        - issuedAt
        - createdAt

    # Admin schemas
...