# lib/api-zod/src/generated/types/createOrderBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Clinic Digital Growth SaaS Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateOrderBodyPlan } from "./createOrderBodyPlan";

export interface CreateOrderBody {
  plan: CreateOrderBodyPlan;
}