# lib/api-zod/src/generated/types/upgradeSubscriptionBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Clinic Digital Growth SaaS Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { UpgradeSubscriptionBodyPlan } from "./upgradeSubscriptionBodyPlan";

export interface UpgradeSubscriptionBody {
  plan: UpgradeSubscriptionBodyPlan;
  durationMonths?: number;
}